rename and move packages
